inspect: Move into internal
