+++ cmd/cometbft/commands/inspect.go
@@ -9,7 +9,7 @@ import (
 	"github.com/spf13/cobra"
 
 	cfg "github.com/cometbft/cometbft/config"
-	"github.com/cometbft/cometbft/inspect"
+	"github.com/cometbft/cometbft/internal/inspect"
 	"github.com/cometbft/cometbft/state"
 	"github.com/cometbft/cometbft/state/indexer/block"
 	"github.com/cometbft/cometbft/store"
