+++ internal/inspect/doc.go
